fix(Updating SSR universal to work):
